+++ pshell/tests/test_call.py
@@ -1,10 +1,20 @@
 import io
+import os
 import tempfile
 import sys
 import time
 import pytest
 import pshell as sh
-from . import StubError
+from . import DATADIR, StubError, unix_only, windows_only
+
+if os.name == 'nt':
+    HELLO_CMD = [os.path.join(DATADIR, 'hello.bat')]
+    EXIT1_CMD = [os.path.join(DATADIR, 'exit1.bat')]
+    SLEEP_CMD = [os.path.join(DATADIR, 'sleep20.bat')]
+else:
+    HELLO_CMD = ['bash', '-c', 'echo Hello world!']
+    EXIT1_CMD = ['bash', '-c', 'exit 1']
+    SLEEP_CMD = ['bash', '-c', 'sleep 2']
 
 
 def test_real_fh_none():
@@ -78,40 +88,47 @@ def test_real_fh_fullpipe():
     assert fh.getvalue() == payload
 
 
+@unix_only
 def test_call():
     assert sh.call('echo "Hello world!" > /dev/null') == 0
 
 
+@unix_only
 def test_call_quotes():
     assert sh.call("echo 'Hello world!' > /dev/null") == 0
 
 
+@unix_only
 def test_call_errexit():
     assert sh.call('notexist.sh') == 127
 
 
+@unix_only
 def test_call_nounset():
     assert sh.call('echo $NOT_EXISTING_VARIABLE') == 1
 
 
+@unix_only
 def test_call_pipefail():
     assert sh.call('cat NOTEXIST | cat') == 1
 
 
+@unix_only
 def test_call_obfuscate_pwd():
     # TODO intercept logging
     assert sh.call('echo -P mypass', obfuscate_pwd='mypass') == 0
 
 
 def test_call_noshell1():
-    assert sh.call(["echo", "Hello world!"], shell=False) == 0
+    assert sh.call(HELLO_CMD, shell=False) == 0
 
 
 def test_call_noshell2():
     with pytest.raises(FileNotFoundError):
-        sh.call('notexist.sh', shell=False)
+        sh.call('notexist', shell=False)
 
 
+@unix_only
 def test_call_timeout():
     ts_start = time.time()
     with pytest.raises(sh.TimeoutExpired):
@@ -119,6 +136,7 @@ def test_call_timeout():
     assert time.time() - ts_start < 0.5
 
 
+@unix_only
 def test_call_real_fh_stringio():
     stderr = io.StringIO()
     stdout = io.StringIO()
@@ -128,53 +146,61 @@ def test_call_real_fh_stringio():
     assert stdout.getvalue() == 'world\n'
 
 
+@unix_only
 def test_call_real_fh_nosetests():
     assert sh.call('echo hello 1>&2 && echo world',
                    stdout=sys.stdout, stderr=sys.stderr) == 0
 
 
+@unix_only
 def test_check_call():
     sh.check_call('echo "Hello world!" > /dev/null')
 
 
+@unix_only
 def test_check_call_quotes():
     sh.check_call("echo 'Hello world!' > /dev/null")
 
 
+@unix_only
 def test_check_call_errexit():
     with pytest.raises(sh.CalledProcessError):
-        sh.check_call('notexist.sh')
+        sh.check_call('notexist')
 
 
+@unix_only
 def test_check_call_nounset():
     with pytest.raises(sh.CalledProcessError):
         sh.check_call('echo $NOT_EXISTING_VARIABLE')
 
 
+@unix_only
 def test_check_call_pipefail():
     with pytest.raises(sh.CalledProcessError):
         sh.check_call('cat NOTEXIST | cat')
 
 
+@unix_only
 def test_check_call_obfuscate_pwd():
     # TODO intercept logging
     sh.check_call('echo -P mypass', obfuscate_pwd='mypass')
 
 
 def test_check_call_noshell1():
-    sh.check_call(["echo", "Hello world!"], shell=False)
+    sh.check_call(HELLO_CMD, shell=False)
 
 
 def test_check_call_noshell2():
     with pytest.raises(sh.CalledProcessError):
-        sh.check_call(['cat', 'notexist.txt'], shell=False)
+        sh.check_call(EXIT1_CMD, shell=False)
 
 
 def test_check_call_noshell3():
     with pytest.raises(FileNotFoundError):
-        sh.check_call('notexist.sh', shell=False)
+        sh.check_call('notexist', shell=False)
 
 
+@unix_only
 def test_check_call_timeout():
     ts_start = time.time()
     with pytest.raises(sh.TimeoutExpired):
@@ -182,6 +208,7 @@ def test_check_call_timeout():
     assert time.time() - ts_start < 0.5
 
 
+@unix_only
 def test_check_call_real_fh_stringio():
     stderr = io.StringIO()
     stdout = io.StringIO()
@@ -191,24 +218,29 @@ def test_check_call_real_fh_stringio():
     assert stdout.getvalue() == 'world\n'
 
 
+@unix_only
 def test_check_call_real_fh_nosetests():
     assert sh.call('echo hello 1>&2 && echo world',
                    stdout=sys.stdout, stderr=sys.stderr) == 0
 
 
+@unix_only
 def test_check_output():
     assert sh.check_output('echo -n "Hello world"') == 'Hello world'
 
 
+@unix_only
 def test_check_output_quotes():
     assert sh.check_output("echo 'Hello world!'") == "Hello world!\n"
 
 
+@unix_only
 def test_check_output_nodecode():
     assert sh.check_output('echo -n "Hello world"',
                            decode=False) == b'Hello world'
 
 
+@unix_only
 def test_check_output_unicode():
     assert sh.check_output(r"printf '\xE2\x98\xA0'") == '☠'
 
@@ -220,58 +252,72 @@ def test_check_output_unicode():
         sh.check_output(r"printf '\x85'", errors='strict')
 
 
+@unix_only
 def test_check_output_errexit():
     with pytest.raises(sh.CalledProcessError):
         sh.check_output('notexist.sh')
 
 
+@unix_only
 def test_check_output_nounset():
     with pytest.raises(sh.CalledProcessError):
         sh.check_output('echo $NOT_EXISTING_VARIABLE')
 
 
+@unix_only
 def test_check_output_pipefail():
     with pytest.raises(sh.CalledProcessError):
         sh.check_output('cat NOTEXIST | cat')
 
 
+@unix_only
 def test_check_output_obfuscate_pwd():
     # TODO intercept logging
     assert sh.check_output('echo -P mypass',
                            obfuscate_pwd='mypass') == "-P mypass\n"
 
 
-def test_check_output_noshell1():
-    assert sh.check_output(['echo', '-n', 'Hello world'],
-                           shell=False) == 'Hello world'
+@unix_only
+def test_check_output_noshell1_unix():
+    assert sh.check_output(HELLO_CMD, shell=False) == 'Hello world!\n'
+
+
+@windows_only
+def test_check_output_noshell1_win():
+    assert sh.check_output(HELLO_CMD, shell=False) \
+        .splitlines()[-1].strip() == 'Hello world!'
 
 
 def test_check_output_noshell2():
     with pytest.raises(sh.CalledProcessError):
-        sh.check_output(['cat', 'notexist.txt'], shell=False)
+        sh.check_output(EXIT1_CMD, shell=False)
 
 
 def test_check_output_noshell3():
     with pytest.raises(FileNotFoundError):
-        sh.check_output('notexist.sh', shell=False)
+        sh.check_output('notexist', shell=False)
 
 
 # Do not change shell=False to True
 # If the timeout expires, the child process will be killed
-# and then waited for again The TimeoutExpired exception will
+# and then waited for again. The TimeoutExpired exception will
 # be re-raised after the child process has terminated.
+# Also the timeout simply does not work in Windows.
+@unix_only
 def test_check_output_timeout():
     ts_start = time.time()
     with pytest.raises(sh.TimeoutExpired):
-        sh.check_output(["sleep", "2"], timeout=0.1, shell=False)
+        sh.check_output(SLEEP_CMD, timeout=0.1, shell=False)
     assert time.time() - ts_start < 0.5
 
 
+@unix_only
 def test_check_output_real_fh_stringio():
     stderr = io.StringIO()
     sh.check_output('echo hello 1>&2', stderr=stderr)
     assert stderr.getvalue() == 'hello\n'
 
 
+@unix_only
 def test_check_output_real_fh_nosetests():
     sh.check_output('echo hello 1>&2', stderr=sys.stderr)