+++ pshell/tests/test_file.py
@@ -88,6 +88,33 @@ def test_remove_noperm(tmpdir):
     assert len(glob.glob(testpath + '/foo.DELETEME.*')) == 1
 
 
+def test_ignore_readonly1(tmpdir):
+    """Test the ignore_readonly=True flag
+    """
+    os.makedirs('%s/foo/bar/baz' % tmpdir)
+    os.chmod('%s/foo/bar/baz' % tmpdir, 0o500)
+    os.chmod('%s/foo/bar' % tmpdir, 0o500)
+    os.chmod('%s/foo' % tmpdir, 0o500)
+
+    with pytest.raises(PermissionError):
+        sh.remove('%s/foo' % tmpdir, recursive=True)
+    assert os.path.exists('%s/foo/bar/baz' % tmpdir)
+
+    sh.remove('%s/foo' % tmpdir, force=False, recursive=True,
+              ignore_readonly=True)
+    assert not os.path.exists('%s/foo' % tmpdir)
+
+
+def test_ignore_readonly2(tmpdir):
+    """Test the case where there was no permission issue to begin with,
+    so a double call to shutil.rmtree would raise FileNotFoundError
+    """
+    os.makedirs('%s/foo/bar' % tmpdir)
+    sh.remove('%s/foo' % tmpdir, force=False, recursive=True,
+              ignore_readonly=True)
+    assert not os.path.exists('%s/foo' % tmpdir)
+
+
 def test_chdir(tmpdir):
     os.environ['UNITTEST_BASH'] = str(tmpdir)
     assert os.getcwd() != str(tmpdir)